# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users/me": {
      "get": {
        "operationId": "UserController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithoutPasswordDto_Output"
                }
              }
            }
          }
        },
        "summary": "Get the current user",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UserController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSelf"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Current user after update",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithoutPasswordDto_Output"
                }
              }
            }
          }
        },
        "summary": "Update the current user",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteSelf"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "User successfully deleted"
          }
        },
        "summary": "Delete the current user",
        "tags": [
          "Users"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UserController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithoutPasswordDto_Output"
                }
              }
            }
          }
        },
        "summary": "Create a new user (Admin only)",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UserController_getAllUsers",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "integer"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched requested users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsers_Output"
                }
              }
            }
          }
        },
        "summary": "Get all users (Admin only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UserController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User_Output"
                }
              }
            }
          }
        },
        "summary": "Get user by ID (Admin only)",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": 0,
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithoutPasswordDto_Output"
                }
              }
            }
          }
        },
        "summary": "Update user by ID (Admin only)",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User successfully deleted"
          }
        },
        "summary": "Delete user by ID (Admin only)",
        "tags": [
          "Users"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_registerUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Auth"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponseDto_Output"
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_loginUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Auth"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse_Output"
                }
              }
            }
          }
        },
        "summary": "Login with email and password",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logoutUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User logged out successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse_Output"
                }
              }
            }
          }
        },
        "summary": "Logout current user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/agent": {
      "get": {
        "operationId": "AgentController_getAgentInfo",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get agent information and configuration",
        "tags": [
          "Agent"
        ]
      }
    },
    "/messages/history": {
      "get": {
        "operationId": "MessagesController_getMessages",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "integer"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get message history",
        "tags": [
          "Messages"
        ]
      }
    },
    "/messages": {
      "post": {
        "operationId": "MessagesController_sendMessage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Send a message to the agent",
        "tags": [
          "Messages"
        ]
      }
    },
    "/messages/stream": {
      "post": {
        "operationId": "MessagesController_sendMessageStream",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Send a message and stream the response",
        "tags": [
          "Messages"
        ]
      }
    },
    "/memories": {
      "get": {
        "operationId": "MemoriesController_listMemoryBlocks",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all memory blocks for the current user",
        "tags": [
          "Memories"
        ]
      },
      "post": {
        "operationId": "MemoriesController_createMemoryBlock",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemoryBlockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a new memory block",
        "tags": [
          "Memories"
        ]
      }
    },
    "/memories/{blockLabel}": {
      "get": {
        "operationId": "MemoriesController_getMemoryBlock",
        "parameters": [
          {
            "name": "blockLabel",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a specific memory block by label",
        "tags": [
          "Memories"
        ]
      },
      "put": {
        "operationId": "MemoriesController_updateMemoryBlock",
        "parameters": [
          {
            "name": "blockLabel",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemoryBlockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a memory block",
        "tags": [
          "Memories"
        ]
      },
      "delete": {
        "operationId": "MemoriesController_deleteMemoryBlock",
        "parameters": [
          {
            "name": "blockLabel",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Memory block successfully deleted"
          }
        },
        "summary": "Delete a memory block",
        "tags": [
          "Memories"
        ]
      }
    }
  },
  "info": {
    "title": "RememberNow API",
    "description": "Epic memory augmentation app!",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "Role_Output": {
        "id": "Role_Output",
        "type": "string",
        "enum": [
          "USER",
          "ADMIN"
        ]
      },
      "UserWithoutPasswordDto_Output": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "role": {
            "default": "USER",
            "$ref": "#/components/schemas/Role_Output"
          },
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "agentId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "role",
          "id",
          "agentId",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "UpdateSelf": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 5,
            "maxLength": 60
          },
          "currentPassword": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          }
        },
        "required": [
          "currentPassword"
        ],
        "id": "UpdateSelf"
      },
      "DeleteSelf": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "confirmationText": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "confirmationText"
        ],
        "id": "DeleteSelf"
      },
      "Role": {
        "id": "Role",
        "type": "string",
        "enum": [
          "USER",
          "ADMIN"
        ]
      },
      "CreateUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 5,
            "maxLength": 60
          },
          "role": {
            "default": "USER",
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "id": "CreateUser"
      },
      "PaginatedUsers_Output": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "maxLength": 254,
                  "format": "email",
                  "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                },
                "role": {
                  "default": "USER",
                  "$ref": "#/components/schemas/Role_Output"
                },
                "id": {
                  "type": "integer",
                  "exclusiveMinimum": 0,
                  "maximum": 9007199254740991
                },
                "agentId": {
                  "type": "string",
                  "nullable": true
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                }
              },
              "required": [
                "email",
                "role",
                "id",
                "agentId",
                "createdAt",
                "updatedAt"
              ],
              "additionalProperties": false
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "limit": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "total": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "totalPages": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "hasNext": {
                "type": "boolean"
              },
              "hasPrev": {
                "type": "boolean"
              }
            },
            "required": [
              "page",
              "limit",
              "total",
              "totalPages",
              "hasNext",
              "hasPrev"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "users",
          "pagination"
        ],
        "id": "PaginatedUsers_Output",
        "additionalProperties": false
      },
      "User_Output": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "role": {
            "default": "USER",
            "$ref": "#/components/schemas/Role_Output"
          },
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "passwordHash": {
            "type": "string"
          },
          "agentId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "role",
          "id",
          "passwordHash",
          "agentId",
          "createdAt",
          "updatedAt"
        ],
        "id": "User_Output",
        "additionalProperties": false
      },
      "UpdateUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 5,
            "maxLength": 60
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "id": "UpdateUser"
      },
      "Auth": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 5,
            "maxLength": 60
          }
        },
        "required": [
          "email",
          "password"
        ],
        "id": "Auth"
      },
      "LoginResponse": {
        "id": "LoginResponse",
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "user": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "maxLength": 254,
                "format": "email",
                "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
              },
              "role": {
                "default": "USER",
                "$ref": "#/components/schemas/Role"
              },
              "id": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "agentId": {
                "type": "string",
                "nullable": true
              },
              "createdAt": {
                "type": "string"
              },
              "updatedAt": {
                "type": "string"
              }
            },
            "required": [
              "email",
              "role",
              "id",
              "agentId",
              "createdAt",
              "updatedAt"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "message",
          "user"
        ],
        "additionalProperties": false
      },
      "SignupResponseDto_Output": {
        "id": "SignupResponse_Output",
        "$ref": "#/components/schemas/LoginResponse"
      },
      "LoginResponse_Output": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "user": {
            "type": "object",
            "properties": {
              "email": {
                "type": "string",
                "maxLength": 254,
                "format": "email",
                "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
              },
              "role": {
                "default": "USER",
                "$ref": "#/components/schemas/Role_Output"
              },
              "id": {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              "agentId": {
                "type": "string",
                "nullable": true
              },
              "createdAt": {
                "type": "string"
              },
              "updatedAt": {
                "type": "string"
              }
            },
            "required": [
              "email",
              "role",
              "id",
              "agentId",
              "createdAt",
              "updatedAt"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "message",
          "user"
        ],
        "id": "LoginResponse_Output",
        "additionalProperties": false
      },
      "LogoutResponse_Output": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "id": "LogoutResponse_Output",
        "additionalProperties": false
      },
      "ChatRequestDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "message"
        ]
      },
      "CreateMemoryBlockDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1
          },
          "value": {
            "default": "",
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "limit": {
            "default": 20000,
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 50000
          },
          "readOnly": {
            "default": false,
            "type": "boolean"
          }
        },
        "required": [
          "label"
        ]
      },
      "UpdateMemoryBlockDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "limit": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "readOnly": {
            "type": "boolean"
          }
        }
      }
    }
  }
}